+++ packages/frontend/src/features/sqlRunner/components/Header.tsx
@@ -1,14 +1,16 @@
 import { ChartKind } from '@lightdash/common';
-import { ActionIcon, Group, Paper, Title, Tooltip } from '@mantine/core';
-import { IconDeviceFloppy, IconLink } from '@tabler/icons-react';
-import { type FC } from 'react';
+import { ActionIcon, Group, Paper, Tooltip } from '@mantine/core';
+import { IconDeviceFloppy } from '@tabler/icons-react';
+import { useCallback, type FC } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import MantineIcon from '../../../components/common/MantineIcon';
+import { EditableText } from '../../../components/VisualizationConfigs/common/EditableText';
 import { useUpdateSqlChartMutation } from '../hooks/useSavedSqlCharts';
 import { useAppDispatch, useAppSelector } from '../store/hooks';
-import { toggleModal } from '../store/sqlRunnerSlice';
+import { DEFAULT_NAME, toggleModal, updateName } from '../store/sqlRunnerSlice';
 import { SaveSqlChartModal } from './SaveSqlChartModal';
+import ShareSqlLinkButton from './ShareSqlLinkButton';
 
 export const Header: FC = () => {
     const { t } = useTranslation();
@@ -28,13 +30,27 @@ export const Header: FC = () => {
         projectUuid,
         savedChartUuid || '',
     );
+
+    const isSaveModalOpen = useAppSelector(
+        (state) => state.sqlRunner.modals.saveChartModal.isOpen,
+    );
+    const onCloseSaveModal = useCallback(() => {
+        dispatch(toggleModal('saveChartModal'));
+    }, [dispatch]);
+
     return (
         <>
-            <Paper shadow="none" radius={0} px="md" py="sm" withBorder>
+            <Paper shadow="none" radius={0} px="md" py="xs" withBorder>
                 <Group position="apart">
-                    <Title order={2} c="gray.6">
-                        {name}
-                    </Title>
+                    <EditableText
+                        size="lg"
+                        placeholder={DEFAULT_NAME}
+                        value={name}
+                        w={400}
+                        onChange={(e) =>
+                            dispatch(updateName(e.currentTarget.value))
+                        }
+                    />
                     <Group spacing="md">
                         <Tooltip
                             variant="xs"
@@ -68,14 +84,16 @@ export const Header: FC = () => {
                             label={t('features_sql_runner_header.share_url')}
                             position="bottom"
                         >
-                            <ActionIcon size="xs">
-                                <MantineIcon icon={IconLink} />
-                            </ActionIcon>
+                            <ShareSqlLinkButton />
                         </Tooltip>
                     </Group>
                 </Group>
             </Paper>
-            <SaveSqlChartModal />
+            <SaveSqlChartModal
+                key={`${isSaveModalOpen}-saveChartModal`}
+                opened={isSaveModalOpen}
+                onClose={onCloseSaveModal}
+            />
         </>
     );
 };